HMM Type

HMM Type

HMMBase.HMM — Method.

HMM(π::Matrix{Float64}, D::Vector{<:Distribution{F}}) where F
HMM(π::Matrix{Float64}, π0::Vector{Float64}, D::Vector{<:Distribution{F}}) where F

Build an HMM with transition matrix π and observations distributions D. If the initial state distribution π0 is not specified, a uniform distribution is assumed.

Observations distributions can be of different types (for example Normal and Exponential). However they must be of the same dimension (all scalars or all multivariates).

Examples

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])

source

HMMBase.assert_hmm — Method.

assert_hmm(π::Matrix{Float64}, π0::Vector{Float64}, D::Vector{<:Distribution})

Throw an AssertionError if the initial state distribution and the transition matrix rows does not sum to 1, and if the observations distributions does not have the same dimensions.

source

HMMBase.sample_hmm — Method.

sample_hmm(hmm::HMM{Univariate}, timesteps::Int)
sample_hmm(hmm::HMM{Multivariate}, timesteps::Int)

Sample a trajectory from hmm. Return a vector of observations for univariate HMMs and a matrix for multivariate HMMs.

source