Inference

Inference

Forward-backward

HMMBase.messages_forwards — Function.

messages_forwards(init_distn::AbstractVector, trans_matrix::AbstractMatrix, log_likelihoods::AbstractMatrix)

Compute forward probabilities, see Forward-backward algorithm.

source

messages_forwards(hmm, observations)

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
z, y = rand(hmm, 1000)
alphas, logtot = messages_forwards(hmm, y)

source

HMMBase.messages_backwards — Function.

messages_backwards(init_distn::AbstractVector, trans_matrix::AbstractMatrix, log_likelihoods::AbstractMatrix)

Compute backward probabilities, see Forward-backward algorithm.

source

messages_backwards(hmm, observations)

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
z, y = rand(hmm, 1000)
betas, logtot = messages_backwards(hmm, y)

source

HMMBase.forward_backward — Function.

forward_backward(init_distn::AbstractVector, trans_matrix::AbstractMatrix, log_likelihoods::AbstractMatrix)

source

forward_backward(hmm, observations)

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
z, y = rand(hmm, 1000)
gammas = forward_backward(hmm, y)

source

Baum–Welch algorithm

HMMBase.fit_mle! — Function.

fit_mle!(hmm::AbstractHMM, observations; eps=1e-3, max_iterations=100, verbose=false)

Perform EM (Baum-Welch) steps until max_iterations is reached, or the change in the log-likelihood is smaller than eps.

Example

hmm, log_likelihood = fit_mle!(hmm, observations)

source

HMMBase.mle_step — Function.

mle_step(hmm::AbstractHMM{F}, observations) where F

Perform one step of the EM (Baum-Welch) algorithm.

Example

hmm, log_likelihood = mle_step(hmm, observations)

source

Viterbi

HMMBase.viterbi — Function.

viterbi(init_distn::AbstractVector, trans_matrix::AbstractMatrix, likelihoods::AbstractMatrix)

Find the most likely hidden state sequence, see Viterbi algorithm.

source

viterbi(trans_matrix::AbstractMatrix, likelihoods::AbstractMatrix)

Assume an uniform initial distribution.

source

viterbi(hmm::HMM, observations::Vector)

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)]);
z, y = rand(hmm, 1000);
z_viterbi = viterbi(hmm, y)
z == z_viterbi

source