Models

HMMBase.AbstractHMM — Type

AbstractHMM{F<:VariateForm}

An HMM type must at-least implement the following interface:

struct CustomHMM{F,T} <: AbstractHMM{F}
    a::AbstractVector{T}               # Initial state distribution
    A::AbstractMatrix{T}               # Transition matrix
    B::AbstractVector{Distribution{F}} # Observations distributions
    # Custom fields ....
end

HMMBase.HMM — Type

HMM([a::AbstractVector{T}, ]A::AbstractMatrix{T}, B::AbstractVector{<:Distribution{F}}) where F where T

Build an HMM with transition matrix A and observations distributions B. If the initial state distribution a is not specified, a uniform distribution is assumed.

Observations distributions can be of different types (for example Normal and Exponential). However they must be of the same dimension (all scalars or all multivariates).

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])

Base.rand — Function

rand(hmm::AbstractHMM, T::Int[, initial_state::Int])

Generate a random trajectory of hmm for T timesteps.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
z, y = rand(hmm, 1000)

rand(hmm::AbstractHMM, z::AbstractVector{Int})

Generate observations from hmm according to trajectory z.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
y = rand(hmm, [1, 1, 2, 2, 1])

HMMBase.permute — Function

permute(hmm::HMM)

Permute the states of hmm according to perm.

Example

hmm = HMM([0.8 0.2; 0.1 0.9], [Normal(0,1), Normal(10,1)])
hmm = permute(hmm, [2, 1])
hmm.A # [0.9 0.1; 0.2 0.8]
hmm.B # [Normal(10,1), Normal(0,1)]

HMMBase.likelihoods — Function

likelihoods(hmm, observations) -> Matrix

Return the likelihood per-state and per-observation.

Arguments

logl::Bool: see common options (TODO).

Output

Matrix{Float64}: a TxK likelihoods matrix.

HMMBase.statdists — Function

statdists(hmm)

Return the stationnary distribution(s) of hmm. That is, the eigenvectors of transpose(hmm.A) with eigenvalues 1.

HMMBase.istransmat — Function

istransmat(A)

Return true if A is square and its rows sums to 1.

HMMBase.nparams — Function

nparams(hmm::AbstractHMM)

Return the number of free parameters in hmm.

NOTE: Does not work, currently, for observations distributions with non-scalar parameters.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
nparams(hmm) # 6

Base.copy — Function

copy(hmm::HMM)

Returns a copy of hmm.

Base.size — Function

size(hmm::AbstractHMM, [dim])

Returns the number of states in the HMM and the dimension of the observations.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
size(hmm) # (2,1)