Types

Types

HMMBase.AbstractHMM — Type.

AbstractHMM{F<:VariateForm}

An HMM type must at-least implement the following interface:

struct CustomHMM{F,T} <: AbstractHMM{F}
    π0::AbstractVector{T}              # Initial state distribution
    π::AbstractMatrix{T}               # Transition matrix
    D::AbstractVector{Distribution{F}} # Observations distributions
    # Custom fields ....
end

source

HMMBase.HMM — Type.

HMM([π0::AbstractVector{T}, ]π::AbstractMatrix{T}, D::AbstractVector{<:Distribution{F}}) where F where T

Build an HMM with transition matrix π and observations distributions D. If the initial state distribution π0 is not specified, a uniform distribution is assumed.

Observations distributions can be of different types (for example Normal and Exponential). However they must be of the same dimension (all scalars or all multivariates).

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])

source

HMMBase.StaticHMM — Type.

StaticHMM([π0::AbstractVector{T}, ]π::AbstractMatrix{T}, D::AbstractVector{<:Distribution{F}}) where {F,T}

See HMM.

source

HMMBase.assert_hmm — Function.

assert_hmm(π0::AbstractVector{Float64}, π::AbstractMatrix{Float64}, D::AbstractVector{<:Distribution})

Throw an AssertionError if the initial state distribution and the transition matrix rows does not sum to 1, and if the observations distributions does not have the same dimensions.

source

Base.size — Method.

size(hmm::AbstractHMM)

Returns the number of states in the HMM and the dimension of the observations.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
size(hmm) # (2,1)

source