Models

HMMBase.HMM — Type

HMM([a, ]A, B) -> HMM

Build an HMM with transition matrix A and observations distributions B. If the initial state distribution a is not specified, a uniform distribution is assumed.

Observations distributions can be of different types (for example Normal and Exponential). However they must be of the same dimension (all scalars or all multivariates).

Arguments

a::AbstractVector{T}: initial probabilities vector.

A::AbstractMatrix{T}: transition matrix.

B::AbstractVector{<:Distribution{F}}: observations distributions.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])

Base.rand — Function

rand([rng, ]hmm, T; ...) -> Matrix | (Vector, Matrix)

Generate a random trajectory of hmm for T timesteps.

Arguments

rng::AbstractRNG (GLOBAL_RNG by default): random number generator to use.

init::Integer (rand(Categorical(hmm.a)) by default): initial state.

seq::Bool (false by default): whether to return the hidden state sequence.

Output

Vector{Int}: hidden state sequence.

Matrix{Float64} (if seq is true): observations (T x dim(obs)).

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
y = rand(hmm, 1000)

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
z, y = rand(hmm, 1000, seq = true)

rand([rng, ]hmm::AbstractHMM, z::AbstractVector{Int}) -> Matrix

Generate observations from hmm according to trajectory z.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
y = rand(hmm, [1, 1, 2, 2, 1])

HMMBase.permute — Function

permute(hmm) -> HMM

Permute the states of hmm according to perm.

Example

hmm = HMM([0.8 0.2; 0.1 0.9], [Normal(0,1), Normal(10,1)])
hmm = permute(hmm, [2, 1])
hmm.A # [0.9 0.1; 0.2 0.8]
hmm.B # [Normal(10,1), Normal(0,1)]

HMMBase.istransmat — Function

istransmat(A) -> Bool

Return true if A is square and its rows sums to 1.

HMMBase.nparams — Function

nparams(hmm) -> Int

Return the number of free parameters in hmm.

Warning

Does not work, currently, for observations distributions with non-scalar parameters.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
nparams(hmm) # 6

Base.copy — Function

copy(hmm) -> HMM

Returns a copy of hmm.

Base.size — Function

size(hmm, [dim]) -> Int | Tuple

Returns the number of states in the HMM and the dimension of the observations.

Example

hmm = HMM([0.9 0.1; 0.1 0.9], [Normal(0,1), Normal(10,1)])
size(hmm) # (2,1)

Likelihoods

HMMBase.likelihoods — Function

likelihoods(hmm, observations) -> Matrix

Return the likelihood per-state and per-observation.

Arguments

logl::Bool: see common options (TODO).

Output

Matrix{Float64}: a TxK likelihoods matrix.

StatsBase.loglikelihood — Function

likelihood(hmm, observations; ...)

Compute the likelihood of the observations under the model. This is defined as the sum of the log of the normalization coefficients in the forward filter.

Arguments

logl, robust: see common options.

Stationnary Distributions

HMMBase.statdists — Function

statdists(hmm)

Return the stationnary distribution(s) of hmm. That is, the eigenvectors of transpose(hmm.A) with eigenvalues 1.

Custom Types

HMMBase.AbstractHMM — Type

AbstractHMM{F<:VariateForm}

A custom HMM type must at-least implement the following interface:

struct CustomHMM{F,T} <: AbstractHMM{F}
    a::AbstractVector{T}               # Initial state distribution
    A::AbstractMatrix{T}               # Transition matrix
    B::AbstractVector{Distribution{F}} # Observations distributions
    # Custom fields ....
end